% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{db_insert_lin}
\alias{db_insert_lin}
\title{\code{db_insert_lin} adds all expression-clinical (linear) edges to a database.}
\usage{
db_insert_lin(expr, clinical, tb_name, db)
}
\arguments{
\item{expr}{matrix of expression levels}

\item{clinical}{matrix of clinical variables (discrete)}

\item{tb_name}{string with name for DB table}

\item{db}{connection to database}
}
\description{
\code{db_insert_lin} adds all expression-clinical (linear) edges to a database.
}
\examples{
library(RSQLite)
mydb <- dbConnect(SQLite(), dbname = "example.db")

db_insert_lin(xx, yy, "graph", mydb)

dbDisconnect(mydb)
}
