% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{db_insert_quad}
\alias{db_insert_quad}
\title{\code{db_insert_quad} adds all expression-expression (quadratic) edges to a database. It inserts edges
using chunks, to limit memory usage.}
\usage{
db_insert_quad(xx, tb_name, chnk_size, db, start = 1, gaussian = FALSE)
}
\arguments{
\item{xx}{matrix of expression levels}

\item{tb_name}{string with name for DB table}

\item{chnk_size}{number of rows in each chunk}

\item{db}{connection to database}

\item{gaussian}{uses gaussian distribution if true; degenerate if false}
}
\description{
\code{db_insert_quad} adds all expression-expression (quadratic) edges to a database. It inserts edges
using chunks, to limit memory usage.
}
