% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools_max_span.R
\name{db_max_span_forest}
\alias{db_max_span_forest}
\title{\code{db_max_span_forest} finds the maximum spanning forest the edges that have been stored in a database.
The function uses a variant of Kruskal's algorithm.Note that it saves a file called \code{forest.rds.backup}.}
\usage{
db_max_span_forest(db, tb_name, n, slices, chunk_size = 100000L,
  verbose = TRUE)
}
\arguments{
\item{db}{connection to a database}

\item{tb_name}{string of table name}

\item{n}{the number of vertices in database}

\item{slices}{the number of slices to compute}

\item{chunk_size}{the size of chunks that will be taken from the database}

\item{verbose}{send output w.r.t. how the process is running}
}
\description{
\code{db_max_span_forest} finds the maximum spanning forest the edges that have been stored in a database.
The function uses a variant of Kruskal's algorithm.Note that it saves a file called \code{forest.rds.backup}.
}
