% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_plotting.R
\name{glasso_plot}
\alias{glasso_plot}
\title{\code{glasso_plot} adds edges to our tree, that are computed
using the glasso algorithm. These makes most sense for vertices
that are roughly Gaussian.}
\usage{
glasso_plot(tree, glasso)
}
\arguments{
\item{tree}{a igraph object of our tree}

\item{rho}{the penalty of the glasso}

\item{dataset}{our dataset to fit the glasso}
}
\description{
\code{glasso_plot} adds edges to our tree, that are computed
using the glasso algorithm. These makes most sense for vertices
that are roughly Gaussian.
}
