% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_plotting.R
\name{make_igraph}
\alias{make_igraph}
\title{\code{make_igraph} makes a igraph object with labelled vertices.}
\usage{
make_igraph(myedges, dataset)
}
\arguments{
\item{dataset}{list of two matrices used for fitting edges}

\item{edges}{edge list of maximum spanning forest.}
}
\description{
\code{make_igraph} makes a igraph object with labelled vertices.
}
