% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_plotting.R
\name{mark_bonferroni_edges}
\alias{mark_bonferroni_edges}
\title{\code{mark_bonferroni_edges} makes a igraph object with labelled vertices.
Edges are coloured w.r.t. the Bonferroni correction with \eqn{\alpha = 0.05}.
This function must be used}
\usage{
mark_bonferroni_edges(g, edges)
}
\arguments{
\item{g}{a igraph object}

\item{edges}{gruph edge list}
}
\value{
a igraph object with pink Bonferroni edges
}
\description{
\code{mark_bonferroni_edges} makes a igraph object with labelled vertices.
Edges are coloured w.r.t. the Bonferroni correction with \eqn{\alpha = 0.05}.
This function must be used
}
