% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_plotting.R
\name{plot_neighbourhood}
\alias{plot_neighbourhood}
\title{\code{plot_neighbourhood} makes a igraph object with labelled vertices.}
\usage{
plot_neighbourhood(g, proband, jumps)
}
\description{
\code{plot_neighbourhood} makes a igraph object with labelled vertices.
}
