% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_plotting.R
\name{plot_neighbourhood_bonferroni}
\alias{plot_neighbourhood_bonferroni}
\title{\code{plot_neighbourhood_bonferroni} makes a igraph object with labelled vertices.
Edges are coloured w.r.t. the Bonferroni correction with \eqn{\alpha = 0.05}.}
\usage{
plot_neighbourhood_bonferroni(g, proband, jumps, edges)
}
\arguments{
\item{g}{a igraph object}

\item{proband}{name of vertex to center neighbourhood around}

\item{jumps}{number of jumps to include in neighbourhood}

\item{edges}{gruph edge list}
}
\description{
\code{plot_neighbourhood_bonferroni} makes a igraph object with labelled vertices.
Edges are coloured w.r.t. the Bonferroni correction with \eqn{\alpha = 0.05}.
}
