% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pval_dataframe}
\alias{pval_dataframe}
\title{\code{pval_dataframe} extract p-values of edges.}
\usage{
pval_dataframe(edges, N)
}
\arguments{
\item{edges}{data frame of edge list}

\item{N}{integer showing the number of observations.}
}
\description{
\code{pval_dataframe} extract p-values of edges.
}
