% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{weight_plot}
\alias{weight_plot}
\title{\code{weight_plot} plots the p-values of weights for model diagnostics.}
\usage{
weight_plot(edges, N)
}
\description{
\code{weight_plot} plots the p-values of weights for model diagnostics.
}
