% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{db_insert_quad}
\alias{db_insert_quad}
\title{\code{db_insert_quad} adds all expression-expression (quadratic) edges to a database. It inserts edges
using chunks, to limit memory usage.}
\usage{
db_insert_quad(xx, tb_name, chnk_size, db, start = 1)
}
\arguments{
\item{xx}{matrix of expression levels}

\item{tb_name}{string with name for DB table}

\item{chnk_size}{number of rows in each chunk}

\item{db}{connection to database}
}
\description{
\code{db_insert_quad} adds all expression-expression (quadratic) edges to a database. It inserts edges
using chunks, to limit memory usage.
}
\examples{
library(RSQLite)
mydb <- dbConnect(SQLite(), dbname = "example.db")
data <- as.matrix(iris[, 1:4])

db_insert_quad(data, "graph", 2, mydb)

dbDisconnect(mydb)
}
