% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{db_max_span_forest}
\alias{db_max_span_forest}
\title{\code{db_max_span_forest} finds the maximum spanning forest the edges that have been stored in a database.
The function uses a variant of Kruskal's algorithm.}
\usage{
db_max_span_forest(db, tb_name, slices)
}
\arguments{
\item{db}{connection to a database}

\item{tb_name}{string of table name}

\item{slices}{the number of slices to compute}
}
\description{
\code{db_max_span_forest} finds the maximum spanning forest the edges that have been stored in a database.
The function uses a variant of Kruskal's algorithm.
}
\examples{
mydb <- dbConnect(SQLite(), dbname = "example.db")
db_max_span_forest(mydb, 100)

}
